change: make `filter` in `resultSelector` optional (#47)
There is certainly the case where we want to e.g. use paging without filtering the data. Then it is quite cumbersome that we have to define a dummy filter which does not filter anything, instead of just omitting the filter property.

Like Paging there is a reasonable default: no filter provided -> no filtering.