add: opensearch query builder - text contains operator (full text search) (#176)
100% of the search terms have to appear in the document.

## Why

We want to be able to search string field which are of type `text` in
the index mapping. So far we could only operate on `keyword` fields. It
should not match in which order the search terms in the `text` appear,
but they should all be present.

## References

- [Match Query
Docs](https://www.elastic.co/guide/en/elasticsearch/reference/7.10/query-dsl-match-query.html)
- Preparation for [VTI-399 - Improve search for CSAF
documents](https://jira.greenbone.net/browse/VTI-399)