remove setting dealing with appending `.keyword` to the field name
For our purposes we can just use the existing query setting `FilterFieldMapping` to map to the actual json path to be used in the query. We don't need different json paths for the same field name depending on the compare operator.

If this ever becomes a usecase we can maybe a `FilterFieldMappingOverride` which can be utilized by the specific CompareOperatorHandler or readd the removed setting. But the default behavior should then be to not further modify the `FilterFieldMapping` (so far we needed to define entries in `WildcardArrays` query setting) .