change: postgres `query` - add mandatory sorting tie breaker setting
Postgres uses the row order stored on disc as tie breakeri (e.g. when no sorting is used or when sorting for a column where different rows have the same value).
The order stored on dics can vary over time and is thus not a good choice. So the user of the query builder has to supply a tie breaker column, ideally one which contains only unique values.