refactor: translate sort direction and logic operator from respective package
There is no gurarantee that the sort package values are valid in a SQL statement. So avoid direct usage to be more robust. Same for the logic operator.

Note: There is nothing meaningful to do for `sorting.NoDirection`. Either Sorting is present, or sorting is omitted. So just fail in this case.