fix: notifications - cache token to avoid auth failure on repeated calls (#273)
Previously we were retrieving the acess token from Keycloak for each
request to the notification service.

This is not efficient and furthermore keycloak deactivates the user
temporarily when it detects an excessing amount of logins.

Moved the authentication to a separate client as it can be useful on its
own in other scenarios. This is a breaking change in the notification
client initialization.